# openapi.yaml
openapi: 3.0.3
info:
  title: Postcredits API
  description: Unified media tracking platform for books and movies with SQL-based recommendations
  version: 1.0.0
  contact:
    name: Team Postcredits
    email: postcredits@tiet.edu.in
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: securePassword123
                displayName:
                  type: string
                  example: John Doe
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
                  email:
                    type: string
        '400':
          description: Invalid input or email already exists
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
                  token:
                    type: string
                  displayName:
                    type: string
        '401':
          description: Invalid credentials
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      operationId: logoutUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user info
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user details
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
                  email:
                    type: string
                  displayName:
                    type: string

  # ==================== Books ====================
  /books:
    get:
      tags:
        - Books
      summary: List all books
      operationId: listBooks
      parameters:
        - name: genre
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
    post:
      tags:
        - Books
      summary: Add a new book (admin only)
      operationId: createBook
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInput'
      responses:
        '201':
          description: Book created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
  /books/{bookId}:
    get:
      tags:
        - Books
      summary: Get book by ID
      operationId: getBook
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Book details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: Book not found

    put:
      tags:
        - Books
      summary: Update book (admin only)
      operationId: updateBook
      security:
        - BearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInput'
      responses:
        '200':
          description: Book updated

    delete:
      tags:
        - Books
      summary: Delete book (admin only)
      operationId: deleteBook
      security:
        - BearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Book deleted

  # ==================== Movies ====================
  /movies:
    get:
      tags:
        - Movies
      summary: List all movies
      operationId: listMovies
      parameters:
        - name: genre
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of movies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movie'
    post:
      tags:
        - Movies
      summary: Add a new movie (admin only)
      operationId: createMovie
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieInput'
      responses:
        '201':
          description: Movie created
  /movies/{movieId}:
    get:
      tags:
        - Movies
      summary: Get movie by ID
      operationId: getMovie
      parameters:
        - name: movieId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Movie details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '404':
          description: Movie not found

    put:
      tags:
        - Movies
      summary: Update movie (admin only)
      operationId: updateMovie
      security:
        - BearerAuth: []
      parameters:
        - name: movieId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieInput'
      responses:
        '200':
          description: Movie updated

    delete:
      tags:
        - Movies
      summary: Delete movie (admin only)
      operationId: deleteMovie
      security:
        - BearerAuth: []
      parameters:
        - name: movieId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Movie deleted

  # ==================== Genres ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Book:
      type: object
      properties:
        workId:
          type: integer
        bookId:
          type: integer
        title:
          type: string
        author:
          type: string
        authorId:
          type: integer
        pages:
          type: integer
        releaseYear:
          type: integer
        genres:
          type: array
          items:
            type: string
        averageRating:
          type: number
        createdAt:
          type: string
          format: date-time
    BookInput:
      type: object
      required:
        - title
        - authorId
        - pages
        - releaseYear
        - genreIds
      properties:
        title:
          type: string
        authorId:
          type: integer
        pages:
          type: integer
        releaseYear:
          type: integer
        genreIds:
          type: array
          items:
            type: integer
    Movie:
      type: object
      properties:
        workId:
          type: integer
        movieId:
          type: integer
        title:
          type: string
        director:
          type: string
        directorId:
          type: integer
        duration:
          type: integer
        releaseYear:
          type: integer
        genres:
          type: array
          items:
            type: string
        averageRating:
          type: number
        createdAt:
          type: string
          format: date-time
    MovieInput:
      type: object
      required:
        - title
        - directorId
        - duration
        - releaseYear
        - genreIds
      properties:
        title:
          type: string
        directorId:
          type: integer
        duration:
          type: integer
        releaseYear:
          type: integer
        genreIds:
          type: array
          items:
            type: integer
...